Remove EDH pages totals api methods & tests
